Campus toggle button being only in the CampusMap page
- Removed the Toggle button from the Outdoor Directions page
- Added it in the CampusMap component and added a function for the toggle to work properly
